explicit cast just in case pthread_t is not of integer type
git-svn-id: svn+ssh://frodo/cluster/svn/projects@11046 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/psc_util/thread.c
@@ -517,7 +517,7 @@ pscthr_getbyid(pthread_t id)
 	psclist_for_each_entry(thr, &psc_threads.pll_listhd,
 	    pscthr_lentry)
 		if (thr->pscthr_pthread == id ||
-		    (unsigned long)thr->pscthr_thrid == id)
+		    (unsigned long)thr->pscthr_thrid == (unsigned long)id)
 			break;
 	PLL_ULOCK(&psc_threads);
 	return (thr);
